Refactor code.py into modules and apps
In order to remove imperative code from code.py, we decided to refactor
the code into modules and apps. The modules are responsible for
interfacing with the hardware, and the apps are responsible for
interfacing with the user. The RetroFrame class is the main class that
is responsible for instantiating the modules and apps.

Each app is a class that inherits from BaseApp. The BaseApp class
provides a common interface for all apps. The most important methods are:
`draw_frame`, and `handle_button_down`.