Bring back width_height_from_bitmap
Unfortunately, if the bitmap does not contain the tile dimensions in the filename, the code was breaking.
dota-diretide.bmp is 32x256, without this code, the height will be set to 256.

Deleting only this function is also inconsistent with `frame_count_from_bitmap` that uses the same logic to count the number of tiles in the tilesheet.

We can't remove the LoopImagesApp yet because of the splash screen.